基本开发完毕

+++ app/src/main/java/com/baimeiyx/www/widget/Ruler.java
@@ -360,7 +360,7 @@ public class Ruler extends View {
 
     }
 
-    public double getDefaultChooseValue() {
-        return currentPosition * valuesInterval / (double) intervalsBetweenValues;
+    public float getDefaultChooseValue() {
+        return currentPosition * valuesInterval / (float) intervalsBetweenValues;
     }
 }
